# Conflicts:
#	Modules/ActiveCampaign/Http/Controllers/ActiveCampaignApiController.php
#	Modules/Trackings/Http/Controllers/TrackingsApiController.php


Former-commit-id: a32008cae6ac2748d39c1c072356b68549a80051

+++ Modules/ActiveCampaign/Http/Controllers/ActiveCampaignApiController.php
@@ -2,19 +2,23 @@
 
 namespace Modules\Activecampaign\Http\Controllers;
 
-use Illuminate\Http\JsonResponse;
-use Illuminate\Routing\Controller;
+
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Log;
-use Modules\Core\Entities\ActivecampaignIntegration;
+use Vinkla\Hashids\Facades\Hashids;
+use Illuminate\Support\Facades\Gate;
 use Modules\Core\Entities\UserProject;
 use Modules\Core\Services\ProjectService;
+use Symfony\Component\HttpFoundation\Response;
 use Modules\Core\Services\ActiveCampaignService;
-use Modules\ActiveCampaign\Transformers\ActivecampaignResource;
+use Modules\Core\Entities\ActivecampaignIntegration;
 use Modules\Projects\Transformers\ProjectsSelectResource;
 use Spatie\Activitylog\Models\Activity;
 use Vinkla\Hashids\Facades\Hashids;
+use Modules\ActiveCampaign\Transformers\ActivecampaignResource;
 
 class ActiveCampaignApiController extends Controller
 {
@@ -43,12 +47,11 @@ class ActiveCampaignApiController extends Controller
                     }
                 }
             }
-
-            return response()->json([
-                                        'integrations' => ActivecampaignResource::collection($activecampaignIntegrations),
-                                        'projects'     => ProjectsSelectResource::collection($projects),
-                                    ]);
-        } catch (Exception $e) {
+            return response()->json(['integrations' => ActivecampaignResource::collection($activecampaignIntegrations),
+                'projects' => ProjectsSelectResource::collection($projects)
+            ]);
+        }
+        catch(Exception $e){
             return response()->json(['message' => 'Ocorreu algum erro'], 400);
         }
     }
@@ -67,6 +70,13 @@ class ActiveCampaignApiController extends Controller
         })->log('Visualizou tela configuração ActiveCampaign');
 
         $activecampaignIntegration = $activecampaignIntegrationModel->find(current(Hashids::decode($id)));
+
+        if (Gate::denies('show', [$activecampaignIntegration])) {
+            return response()->json([
+                    'message' => 'Sem permissão',
+                ],Response::HTTP_FORBIDDEN
+            );
+        }
 
         return new ActivecampaignResource($activecampaignIntegration);
     }
@@ -160,6 +170,13 @@ class ActiveCampaignApiController extends Controller
                 $projectId   = current(Hashids::decode($id));
                 $integration = $activecampaignIntegrationModel->where('project_id', $projectId)->first();
 
+                if (Gate::denies('edit', [$activecampaignIntegration])) {
+                    return response()->json([
+                            'message' => 'Sem permissão',
+                        ],Response::HTTP_FORBIDDEN
+                    );
+                }
+
                 if ($integration) {
                     return response()->json(['projects' => $projects, 'integration' => $integration]);
                 } else {
@@ -195,6 +212,14 @@ class ActiveCampaignApiController extends Controller
             $data                           = $request->all();
             $integrationId                  = current(Hashids::decode($id));
             $activecampaignIntegration      = $activecampaignIntegrationModel->find($integrationId);
+
+
+            if (Gate::denies('update', [$activecampaignIntegration])) {
+                return response()->json([
+                        'message' => 'Sem permissão',
+                    ],Response::HTTP_FORBIDDEN
+                );
+            }
 
             if (empty($data['api_url']) || empty($data['api_key'])) {
                 return response()->json([
@@ -245,7 +270,16 @@ class ActiveCampaignApiController extends Controller
             $integrationId                  = current(Hashids::decode($id));
             $activecampaignIntegrationModel = new ActivecampaignIntegration();
             $integration                    = $activecampaignIntegrationModel->find($integrationId);
-            $integrationDeleted             = $integration->delete();
+
+            if (Gate::denies('destroy', [$integration])) {
+                return response()->json([
+                        'message' => 'Sem permissão',
+                    ],Response::HTTP_FORBIDDEN
+                );
+            }
+
+            $integrationDeleted = $integration->delete();
+
             if ($integrationDeleted) {
                 return response()->json([
                                             'message' => 'Integração Removida com sucesso!',
